cleans up python scripts and makes them also run with python3
